backend-ai/pyproject.toml
[tool.poetry]
name = "backend-ai"
version = "1.0.0"
authors = ["Anton Gorokh"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
openai = "^1.77.0"
python-dotenv = "^1.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "main:start"